requirements.txt
streamlit
pandas
langchain
Pillow
transformers
